% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplineFit.R
\name{fitSpline}
\alias{fitSpline}
\title{Fitting the optimal spline by cross validation method.}
\usage{
fitSpline(x, y, splineDegree, numKnots)
}
\arguments{
\item{x}{independent variable values.}

\item{y}{dependent variable values}
}
\value{
A list containing fitted values, optimal degree and number of knots, coefficients.
}
\description{
This function fits an optimal spline function to the given data set after choosing optimal degree of the polynomial and optimal number of knots by cross-validation.
}
