% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPE.R
\name{MSPE_boot}
\alias{MSPE_boot}
\title{Bootstrap MSPE Estimator of OBP}
\usage{
MSPE_boot(theta.OBP, D, x, L = 200)
}
\arguments{
\item{theta.OBP}{A vector of OBPs.}

\item{D}{Scalar, Vector (or diagonal matrix) of error variances.}

\item{L}{is the number of bootstrap samples. the default is 200.}
}
\value{
This function will return the Bootstrap MSPE estimator of OBP.
}
\description{
This function computes the bootstrap MSPE estimator of OBP.
}
\references{
Jiang J, Nguyen T, and Rao J. S. (2011), "Best Predictive Small Area Estimation", Journal of the American Statistical Association.
}
