% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPE.R
\name{MSPE_boot_augmented}
\alias{MSPE_boot_augmented}
\title{Bootstrap MSPE estimator of Augmented Observed Best Predictor (Augmented OBP).}
\usage{
MSPE_boot_augmented(theta.OBP.augmented, D, x, weight, L = 200)
}
\arguments{
\item{theta.OBP.augmented}{A vector of augmented OBPs.}

\item{D}{Scalar, Vector (or diagonal matrix) of random error variances.}

\item{x}{Design matrix without intercept term.}

\item{weight}{Weight of each small area for calculating the benchmarked OBP (generally sampling weights of each small area).}

\item{L}{Number of bootstrap samples. the default is 200.}
}
\value{
This function will return a vector of the Bootstrap MSPE estimator of Augmented OBP.
}
\description{
This function computes the bootstrap MSPE estimator of benchmarked augmented OBP.
}
\references{
Bandyopadhyay R, Jiang J (2017) "Benchmarking the Observed Best Predictor"

Jiang J, Nguyen T, and Rao J. S. (2011), "Best Predictive Small Area Estimation", Journal of the American Statistical Association.
}
