% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPE.R
\name{MSPE_JNR}
\alias{MSPE_JNR}
\title{JNR MSPE Estimator of OBP}
\usage{
MSPE_JNR(formula, data, errorvar, theta.OBP, A.BPE, beta.BPE)
}
\arguments{
\item{theta.OBP}{A vector of OBP values}

\item{A.BPE}{BPE estimate of variance of random effects.}

\item{y}{A vector of the response variable.}

\item{D}{Scalar, Vector (or diagonal matrix) of error variances.}
}
\value{
This function will return the Naive MSPE estimator of the OBP.
}
\description{
This function computes the JNR MSPE estimator of OBP.
}
\references{
Jiang J, Nguyen T, and Rao J. S. (2011), "Best Predictive Small Area Estimation", Journal of the American Statistical Association.
}
