% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MSPE2.R
\name{MSPE_McJack}
\alias{MSPE_McJack}
\title{McJack MSPE estimator}
\usage{
MSPE_McJack(formula, data, errorvar, A.BPE, K = 1000, returnMcSpline = TRUE)
}
\arguments{
\item{formula}{an object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. The variables included in formula must have a length equal to the number of small areas. More about the model specification are given under Details.}

\item{data}{optional data frame containing the variable names in \code{formula}.}

\item{errorvar}{vector containing the variances of the random error for each small area.}

\item{A.BPE}{optional BPE estimate of random effects variance.}

\item{K}{number of Monte Carlo simulations. Default is 1000.}

\item{returnMcSpline}{logical. Returns McSpline estimator with McJack (default).}
}
\value{
The function will return a list with the following objects.
\item{McJack}{ McJack estimator of the MSPE of OBP.}
\item{McSpline}{ McSpline estimator of the MSPE of OBP. This is returned by default. To turn this feature off, set \code{returnMcSpline = FALSE}.}
}
\description{
This function computes the McJack MSPE estimator of the Observed Best Predictor (OBP).
}
\details{
\code{formula} is specified in the form \code{response ~ predictor} where the predictor is univariate. \code{formula} has an implied intercept term. To remove the intercept term, use either \code{y ~ x - 1} or \code{y ~ 0 + x}.\cr
\cr If \code{A.BPE} is missing, the function computes the BPE from data. User can also provide the true A instead if that is known.
}
\references{
Bandyopadhyay R, Jiang J (2017) "Benchmarking the Observed Best Predictor"

Jiang J, Nguyen T, and Rao J. S. (2011), "Best Predictive Small Area Estimation", Journal of the American Statistical Association.
}

