% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPE_Benchmark.R
\name{MSPE_McJack_Benchmark}
\alias{MSPE_McJack_Benchmark}
\title{Mc Jack MSPE benchmark estimator}
\usage{
MSPE_McJack_Benchmark(formula, data, errorvar, weight, A.BPE, K = 1000,
  returnMcSpline = TRUE)
}
\arguments{
\item{formula}{an object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. The variables included in formula must have a length equal to the number of domains D. Details of model specification are given under Details.}

\item{data}{data frame containing the variable names in formula and errorvar.}

\item{errorvar}{vector containing the D sampling variances of direct estimators for each domain. The values must be sorted as the variables in formula.}

\item{weight}{vector of weights.}

\item{A.BPE}{BPE estimator of random error variance component. Recomputes if missing.}

\item{K}{number of Monte Carlo simulations. Default is 1000.}

\item{returnMcSpline}{logical. Returns McSpline MSPE benchmark estimators with McJack (default).}
}
\value{
The function will return a list with the McJack and McSpline (if returnMcSpline is TRUE) MSPE benchmark estimators, both adjusted and augmented methods.
}
\description{
This function computes the McJack MSPE benchmark estimator.
}
\references{
Bandyopadhyay R, Jiang J (2017) "Benchmarking of Observed Best Predictor"

Jiang J, Nguyen T, and Rao J. S. (2011), "Best Predictive Small Area Estimation", Journal of the American Statistical Association.
}
