% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MSPE2.R
\name{MSPE_McSpline}
\alias{MSPE_McSpline}
\title{Mc Spline MSPE estimator}
\usage{
MSPE_McSpline(formula, data, errorvar, A.BPE, K = 1000)
}
\arguments{
\item{formula}{an object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. The variables included in formula must have a length equal to the number of domains D. Details of model specification are given under Details.}

\item{data}{data frame containing the variable names in formula and errorvar.}

\item{errorvar}{vector containing the D sampling variances of direct estimators for each domain. The values must be sorted as the variables in formula.}

\item{A.BPE}{BPE estimator of random error variance component. Recomputes if missing.}

\item{K}{number of Monte Carlo simulations. Default is 1000.}
}
\value{
The function will return a list with the McSpline MSPE estimator.
}
\description{
This function computes the McSPline MSPE estimator of Observed Best Predictor(OBP).
}
\references{
Bandyopadhyay R, Jiang J (2017) "Benchmarking of Observed Best Predictor"

Jiang J, Nguyen T, and Rao J. S. (2011), "Best Predictive Small Area Estimation", Journal of the American Statistical Association.
}

