% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obpFH.R
\name{bpeFH}
\alias{bpeFH}
\title{Best predictive estimator for Fay-Herriot model.}
\usage{
bpeFH(formula, data, errorvar, randvar = NULL, maxiter = 100,
  precision = 1e-04)
}
\arguments{
\item{formula}{an object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. The variables included in formula must have a length equal to the number of small areas. Details of model specification are given under Details.}

\item{data}{optional data frame containing the variable names in \code{formula}.}

\item{errorvar}{vector containing the variances of the random error for each small area.}

\item{randvar}{varinace of the random effect. If not supplied, BPE is estimated.}

\item{maxiter}{maximum number of iterations used in estimating randvar.}

\item{precision}{covergence tolerance limit for estimating randvar.}
}
\value{
The function will return a list with the following objects.
\item{A.BPE}{BPE of variance of the random effect (if not specified by the user).}
\item{beta.BPE}{BPE of the regression coefficient of the fixed effects.}
}
\description{
This function computes the best predictive estimators (BPE) of the unknown parameters for Fay-Herriot model.
}
\details{
\code{formula} is specified in the form \code{response ~ predictors} where the predictors are separated by \code{+}. \code{formula} has an implied intercept term. To remove the intercept term, use either \code{y ~ x - 1} or \code{y ~ 0 + x}.
}
\references{
Jiang J, Nguyen T, and Rao J. S. (2011), "Best Predictive Small Area Estimation", Journal of the American Statistical Association.
}

