% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplineFit.R
\name{fitSpline.cv}
\alias{fitSpline.cv}
\title{Fitting the optimal spline by cross validation method.}
\usage{
fitSpline.cv(x, y)
}
\arguments{
\item{x}{independent variable values.}

\item{y}{dependent variable values}
}
\value{
A list containing the following objects.
\item{f.hat}{ fitted values.}
\item{p}{ degree of the fitted polynomial.}
\item{q}{ number of knots.}
\item{beta}{ estimated coefficients of the polynomial.}
\item{gamma}{ estimated coefficients of the truncated polynomial.}
\item{knots}{ location of the knots.}
}
\description{
This function fits an optimal spline function to the given data set after choosing optimal degree of the polynomial and optimal number of knots by cross-validation.
}
\details{
The equation of a spline with degree \eqn{p} and \eqn{q} knots \eqn{\xi_1, \xi_2, \cdots, \xi_q}{\xi_1, \xi_2, ..., \xi_q} is
\deqn{y = \beta_0 + \sum^p_{k=1}\beta_k x^k + \sum^q_{l=1}\gamma_l(x-\xi_l)^p_+ .}{y = \beta_0 + \beta_1 * x + \beta_2 * x^2 + ... + \beta_p * x^p + \gamma_1 * (pos(x - \xi_1))^p + ... + \gamma_q * (pos(x - \xi_q))^p}
where \eqn{u_+ = u I(u > 0)}{pos(u) = u * I(u > 0)}.
}
