% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Income}
\alias{Income}
\title{Income data}
\format{A data frame with NNNN rows and NN variables:
\describe{
  \item{subject}{Anonymized Mechanical Turk Worker ID}
  \item{trial}{Trial number, from 1..NNN}
  ...
}}
\usage{
Income
}
\description{
This is data from the first experiment ever to try XYZ using Mechanical
Turk workers.
}
\keyword{datasets}
