% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPE.R
\name{MSPE_boot_augmented}
\alias{MSPE_boot_augmented}
\title{Bootstrap MSPE augmented estimator of OBP}
\usage{
MSPE_boot_augmented(theta.OBP.augmented, D, x, weight, L = 200)
}
\arguments{
\item{D}{Scalar, Vector (or diagonal matrix) of error variances.}

\item{L}{is the number of bootstrap samples. the default is 200.}

\item{theta.OBP}{A vector of OBPs.}
}
\value{
This function will return the Bootstrap MSPE estimator of OBP.
}
\description{
This function computes the bootstrap MSPE estimatr of OBP.
}
\references{
Jiang J, Nguyen T, and Rao J. S. (2011), "Best Predictive Small Area Estimation", Journal of the American Statistical Association.
}
