% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSPE_Benchmark.R
\name{MSPE_McJack_Benchmark}
\alias{MSPE_McJack_Benchmark}
\title{McJack MSPE estimator of benchmarked OBP}
\usage{
MSPE_McJack_Benchmark(formula, data, errorvar, weight, A.BPE, K = 1000,
  returnMcSpline = TRUE)
}
\arguments{
\item{formula}{an object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. The variables included in formula must have a length equal to the number of small areas. More about the model specification are given under Details.}

\item{data}{optional data frame containing the variable names in \code{formula}.}

\item{errorvar}{vector containing the variances of the random errors for all small areas.}

\item{weight}{vector containing the sampling weights of small areas. If sum of the weights is not 1, the weights are normalized.}

\item{A.BPE}{optional BPE estimate of random effects variance.}

\item{K}{number of Monte Carlo simulations. Default is 1000.}

\item{returnMcSpline}{logical. Returns McSpline MSPE benchmark estimators with McJack (default).}
}
\value{
The function will return a list with the following objects.
\item{McJack_Adj_bench}{McJack MSPE estimator of the adjusted OBP.}
\item{McJack_Aug_bench}{McJack MSPE estimator of the augmented OBP.}
\item{McSpline_Adj_bench}{McSpline MSPE estimator of the adjusted OBP. This is returned by default. To turn this feature off, set \code{returnMcSpline = FALSE}}
\item{McSpline_Aug_bench}{McSpline MSPE estimator of the augmented OBP. This is returned by default. To turn this feature off, set \code{returnMcSpline = FALSE}}
}
\description{
This function computes the McJack MSPE estimator of the benchmarked Observed Best Predictor.
}
\details{
\code{formula} is specified in the form \code{response ~ predictor} where the predictor is univariate. \code{formula} has an implied intercept term. To remove the intercept term, use either \code{y ~ x - 1} or \code{y ~ 0 + x}.\cr
\cr If \code{A.BPE} is missing, the function computes the BPE from data. User can also provide the true A instead if that is known.
}
\references{
Bandyopadhyay R, Jiang J (2017) "Benchmarking the Observed Best Predictor"

Jiang J, Nguyen T, and Rao J. S. (2011), "Best Predictive Small Area Estimation", Journal of the American Statistical Association.
}
