% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obpFHbenchmark.R
\name{obpFH_adjusted}
\alias{obpFH_adjusted}
\title{Adjusted observed best predictor satisfying benchmark condition for Fay-Herriot model.}
\usage{
obpFH_adjusted(formula, data, errorvar, weight, randvar = NULL,
  maxiter = 100, precision = 1e-04)
}
\arguments{
\item{formula}{an object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. The variables included in formula must have a length equal to the number of domains D. Details of model specification are given under Details.}

\item{data}{data frame containing the variable names in formula and errorvar.}

\item{errorvar}{vector containing the D sampling variances of direct estimators for each domain. The values must be sorted as the variables in formula.}

\item{weight}{vector containing the sampling weights of small areas. Default is uniform. If sum of the weights is not 1, the weights are normalized.}

\item{randvar}{varinace of the random effect. If not supplied, BPE is estimated.}

\item{maxiter}{maximum number of iterations used in estimating randvar.}

\item{precision}{covergence tolerance limit for estimating randvar.}

\item{method}{string specifying the benchmarking method. Options are "adjusted" and "augmented". Computes both if not specified. See references for details.}
}
\value{
The function will return a list consisting of the OBP of the small area mean, BPE of the regression coefficient of the fixed effect and BPE of variance of the random effect (if not specified by the user).
}
\description{
This function computes the observed best predictor (OBP) for Fay-Herriot model. The variance of the random error can be specified by the user. Otherwise the function will calculate its Best Predictive Estimator (BPE). In the process of of computing OBP it also calculates the BPE of the regression coefficients of the fixed effect
}
\references{
Bandyopadhyay R, Jiang J (2017) "Benchmarking of Observed Best Predictor"

Jiang J, Nguyen T, and Rao J. S. (2011), "Best Predictive Small Area Estimation", Journal of the American Statistical Association.
}
